modified frontend script

+++ Shell_scripts/common.sh
@@ -23,9 +23,17 @@ systemd_start () {
 }
 
 nginx_setup () {
-rm -rf /usr/share/nginx/html/*
-curl -o /tmp/frontend.zip https://roboshop-artifacts.s3.amazonaws.com/frontend-v3.zip
-cd /usr/share/nginx/html
-unzip /tmp/frontend.zip
-cp $PWD/nginx.conf /etc/nginx/nginx.conf
+ls /usr/share/nginx/html/*
+ if [ $? -eq 0 ]; then
+     curl -o /tmp/frontend.zip https://roboshop-artifacts.s3.amazonaws.com/frontend-v3.zip
+     cd /usr/share/nginx/html
+     unzip /tmp/frontend.zip
+     cp $PWD/nginx.conf /etc/nginx/nginx.conf
+ else
+     rm -rf /usr/share/nginx/html/*
+     curl -o /tmp/frontend.zip https://roboshop-artifacts.s3.amazonaws.com/frontend-v3.zip
+     cd /usr/share/nginx/html
+     unzip /tmp/frontend.zip
+     cp $PWD/nginx.conf /etc/nginx/nginx.conf
+fi
 }
